requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
google-cloud-ndb==2.2.2
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2
itsdangerous==2.1.2
gunicorn==21.2.0
PyJWT==2.8.0
